Fixed some compiler warnings from VS2013

+++ src/initialization/FGSimplexTrim.cpp
@@ -40,7 +40,7 @@ FGSimplexTrim::FGSimplexTrim(FGFDMExec * fdm, TrimMode mode)
     double abstol = node->GetDouble("trim/solver/abstol");
     double speed = node->GetDouble("trim/solver/speed"); // must be > 1, 2 typical
     double random = node->GetDouble("trim/solver/random");
-    int iterMax = node->GetDouble("trim/solver/iterMax");
+    int iterMax = node->GetInt("trim/solver/iterMax");
     bool showConvergence = node->GetBool("trim/solver/showConvergence");
     bool pause = node->GetBool("trim/solver/pause");
     bool showSimplex = node->GetBool("trim/solver/showSimplex");
